Uses a totalItems() function that calculates size of the cache when called rather than trying to track the total size manually

+++ lib/srv/desktop/windows_server_test.go
@@ -371,7 +371,7 @@ func TestAuditCacheLifecycle(t *testing.T) {
 
 	// Confirm that audit cache entry for sid
 	// is in the expected state.
-	require.Equal(t, uint32(1), entry.totalItems)
+	require.Equal(t, 1, entry.totalItems())
 	name, ok := s.auditCache.GetName(sessionID(sid), directoryID(did))
 	require.True(t, ok)
 	require.Equal(t, directoryName(testDirName), name)
@@ -391,7 +391,7 @@ func TestAuditCacheLifecycle(t *testing.T) {
 	encoded, err := readReq.Encode()
 	require.NoError(t, err)
 	sendHandler(readReq, encoded)
-	require.Equal(t, uint32(2), entry.totalItems)
+	require.Equal(t, 2, entry.totalItems())
 
 	// A SharedDirectoryWriteRequest should add a corresponding entry in the writeRequestCache.
 	writeReq := tdp.SharedDirectoryWriteRequest{
@@ -404,7 +404,7 @@ func TestAuditCacheLifecycle(t *testing.T) {
 	encoded, err = writeReq.Encode()
 	require.NoError(t, err)
 	sendHandler(writeReq, encoded)
-	require.Equal(t, uint32(3), entry.totalItems)
+	require.Equal(t, 3, entry.totalItems())
 
 	// Check that the readRequestCache was properly filled out.
 	_, ok = entry.readRequestCache[completionID(cid)]
@@ -422,7 +422,7 @@ func TestAuditCacheLifecycle(t *testing.T) {
 		ReadData:       []byte{}, // irrelevant in this context
 	}
 	recvHandler(readRes)
-	require.Equal(t, uint32(2), entry.totalItems)
+	require.Equal(t, 2, entry.totalItems())
 
 	// SharedDirectoryWriteResponse should cause the entry in the writeRequestCache to be cleaned up.
 	writeRes := tdp.SharedDirectoryWriteResponse{
@@ -431,7 +431,7 @@ func TestAuditCacheLifecycle(t *testing.T) {
 		BytesWritten: length,
 	}
 	recvHandler(writeRes)
-	require.Equal(t, uint32(1), entry.totalItems)
+	require.Equal(t, 1, entry.totalItems())
 
 	// Check that the readRequestCache was properly cleaned up.
 	_, ok = entry.readRequestCache[completionID(cid)]
